+++ app/[locale]/job/[idx]/expense/snap/page.test.tsx
