+++ app/[locale]/job/[idx]/expense/snap/page.tsx
