+++ app/[locale]/job/[idx]/expense/snap/recommend/page.tsx
